requirements.txt
apache-airflow==2.1.4
psycopg2-binary==2.9.1
sqlalchemy<2.0
apache-airflow[postgres]